Fix register read/write and calculations on AVR int is 16 bits long on this architecture

+++ TMC5160.cpp
@@ -325,7 +325,7 @@ bool TMC5160::setEncoderResolution(int motorSteps, int encResolution, bool inver
 #endif
 
 		//Check if the decimal prescaler gives an exact match. Floats have about 7 digits of precision so no worries here.
-		return ((int)(factor * 10000.0f) * encResolution == motorSteps * _uStepCount * 10000);
+		return ((int)(factor * 10000.0f) * encResolution == motorSteps * (int)_uStepCount * 10000);
 	}
 }
